# swagger.json
{"openapi":"3.0.0","paths":{"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/user/self":{"get":{"operationId":"UserController_findSelf","parameters":[],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/post":{"post":{"operationId":"PostController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostDto"}}}},"responses":{"201":{"description":""}},"tags":["post"],"security":[{"bearer":[]}]},"get":{"operationId":"PostController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["post"],"security":[{"bearer":[]}]}},"/post/{id}":{"get":{"operationId":"PostController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["post"],"security":[{"bearer":[]}]},"patch":{"operationId":"PostController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostDto"}}}},"responses":{"200":{"description":""}},"tags":["post"],"security":[{"bearer":[]}]},"delete":{"operationId":"PostController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["post"],"security":[{"bearer":[]}]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocalLoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/sign-up":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}}},"info":{"title":"All In One API","description":"All In One API 입니다","version":"1.0","contact":{}},"tags":[{"name":"all-in-one","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"username":{"type":"string"},"nickname":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"email":{"type":"string"}},"required":["username","nickname","roles","email"]},"UpdateUserDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"nickname":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"email":{"type":"string"}},"required":["id","username","nickname","roles","email"]},"CreatePostDto":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"authorId":{"type":"string"}},"required":["title","content","authorId"]},"UpdatePostDto":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"content":{"type":"string"},"authorId":{"type":"string"}},"required":["id","title","content","authorId"]},"LocalLoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]}}}}